Corrected count API of Upload paid status.

+++ config/src/main/java/com/gl/ceir/config/configuration/PropertiesReader.java
@@ -13,5 +13,5 @@ public class PropertiesReader {
 	public String dateViewFormat;
 	
 	@Value("${default-no-of-regularized-devices}")
-	public int defaultNoOfRegularizedDevices;
+	public Long defaultNoOfRegularizedDevices;
 }
